Avoid importing vite in db utils
