+++ packages/db/src/core/utils.ts
@@ -1,6 +1,5 @@
-import type { AstroConfig, AstroIntegration } from 'astro';
+import type { AstroConfig } from 'astro';
 import { loadEnv } from 'vite';
-import type { AstroDbIntegration } from './types.js';
 
 export type VitePlugin = Required<AstroConfig['vite']>['plugins'][number];
 
@@ -28,21 +27,4 @@ export function getDbDirectoryUrl(root: URL | string) {
 	return new URL('db/', root);
 }
 
-export function defineDbIntegration(integration: AstroDbIntegration): AstroIntegration {
-	return integration;
-}
-
 export type Result<T> = { success: true; data: T } | { success: false; data: unknown };
-
-/**
- * Map an object's values to a new set of values
- * while preserving types.
- */
-export function mapObject<T, U = T>(
-	item: Record<string, T>,
-	callback: (key: string, value: T) => U
-): Record<string, U> {
-	return Object.fromEntries(
-		Object.entries(item).map(([key, value]) => [key, callback(key, value)])
-	);
-}
