+++ packages/db/src/utils.ts
@@ -1,7 +1,11 @@
-export { defineDbIntegration } from './core/utils.js';
+import type { AstroIntegration } from 'astro';
 import { tableSchema } from './core/schemas.js';
-import type { ColumnsConfig, TableConfig } from './core/types.js';
+import type { AstroDbIntegration, ColumnsConfig, TableConfig } from './core/types.js';
 import { type Table, asDrizzleTable as internal_asDrizzleTable } from './runtime/index.js';
+
+export function defineDbIntegration(integration: AstroDbIntegration): AstroIntegration {
+	return integration;
+}
 
 export function asDrizzleTable<
 	TableName extends string = string,
